UI: Hide AppState within UiCtrlImpl, and change it to store controllers.
AppState doesn't need to be exposed in the public interface of UiCtrl.  It can
just be part of UiCtrlImpl, as there's no need to have MainWindow managing open
window counts.  Also, we want to be tracking the controllers themselves, not
just the count of running controllers, so that we can have coordination between
controllers, e.g. a "File > Quit" that asks to save all open games before
closing.