UI: Fix navigating between nodes with comments causing dirtyness.
When a node and a direct child of it both contain comments, then navigating
between the two in the UI triggers the UI dirty state due to the
model->view copy updating the contents of a GTK+ TextBuffer, which
replaces its text by 1) removing the old contents, and 2) inserting the new
contents.  Except, the "text buffer changed" signal fires after both 1) and
2), causing us to write an empty value back to the comment property for the
target node, triggering dirtiness because the property has changed.

The changes to Monad and MonadTest aren't part of the bug fix.  They're
just minor improvements.