Make RealValue use SgfValue: a custom, exact bigfloat implementation.
This should resolve the issues with needing to go through Double to render
and parse Rationals; namely, unrepresentable numbers and limited precision.
Although we're not likely to be working with highly precise numbers, we
should be able to represent something like 0.1 accurately.  Also, sometimes
sometimes 'show' returns scientific notation for doubles, and we don't want
that.