Fix exports and improve documentation in the Property module.
Documentation of modules reexported in Game.Goatee.Lib.Property was broken
in 82bca10.  This fixes that by changing the reexport style.  Also, we add
documentation to things exported from Property, hide things that shouldn't
be exported, and split Game.Goatee.Lib.Property.Parser off as a separate
import (it's not as generally useful as the others).

Fixes: https://savannah.nongnu.org/bugs/?43077